refactor: extension shortcut feature
