+++ packages/shared/src/components/modals/common.tsx
@@ -267,6 +267,20 @@ const ContentModal = dynamic(
   () => import(/* webpackChunkName: "contentModal" */ './ContentModal'),
 );
 
+const MostVisitedSitesModal = dynamic(
+  () =>
+    import(
+      /* webpackChunkName: "mostVisitedSitesModal" */ './shortcuts/MostVisitedSitesModal'
+    ),
+);
+
+const CustomLinksModal = dynamic(
+  () =>
+    import(
+      /* webpackChunkName: "customLinksModal" */ './shortcuts/CustomLinksModal'
+    ),
+);
+
 export const modals = {
   [LazyModal.SquadMember]: SquadMemberModal,
   [LazyModal.UpvotedPopup]: UpvotedPopupModal,
@@ -312,6 +326,8 @@ export const modals = {
   [LazyModal.GiftPlusReceived]: GiftReceivedPlusModal,
   [LazyModal.GiveAward]: GiveAwardModal,
   [LazyModal.ContentModal]: ContentModal,
+  [LazyModal.MostVisitedSites]: MostVisitedSitesModal,
+  [LazyModal.CustomLinks]: CustomLinksModal,
 };
 
 type GetComponentProps<T> = T extends
